% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcpt.R
\name{factor_model}
\alias{factor_model}
\title{Fit a factor_model}
\usage{
factor_model(
  x,
  max.q = NULL,
  q = NULL,
  bn = TRUE,
  bn.op = 2,
  normalisation = TRUE
)
}
\arguments{
\item{x}{data matrix, with series as columns}

\item{max.q}{integer maximum factor number to consider}

\item{q}{integer factor number}

\item{bn}{Boolean, use information criteria of Bai/Ng (2002) for estimating q}

\item{bn.op}{integer from 1 to 5 determining which information criterion to use (see details)}

\item{normalisation}{Boolean, centre and scale x}
}
\value{
factor_model object
}
\description{
Extracts factors, loadings, and estimates the factor number
}
\examples{
factor_model(panel$panel)
}
