% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get entire nowcasting dataset}
\usage{
get_data(observation_end = "2021-06-01")
}
\arguments{
\item{observation_end}{Date denoting end of period, of format YYYY-MM-01}
}
\value{
Balanced panel of nowcasting dataset
}
\description{
Get entire nowcasting dataset
}
\examples{
nowcasting_data <- get_data()
}
