% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumvar_factor.R
\name{mosumvar_factor}
\alias{mosumvar_factor}
\title{Change point detection for a factor model with VAR dynamics}
\usage{
mosumvar_factor(
  x,
  p,
  G,
  method = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25,
  do_bootstrap = FALSE,
  M = 1000,
  univ = FALSE,
  rm_cross_terms = TRUE,
  global_resids = TRUE,
  r = NULL,
  r_max = NA
)
}
\arguments{
\item{x}{matrix of data with series as columns}

\item{p}{integer VAR order}

\item{G}{integer MOSUM bandwidth}

\item{method}{String, which of "Wald" or "Score" statistics to use}

\item{estim}{String, estimation method}

\item{varEstim}{String, variance estimation method}

\item{alpha}{numeric, significance level}

\item{criterion}{String, location criterion}

\item{nu}{numeric, location criterion parameter}

\item{do_bootstrap}{Boolean, perform threshold bootstrap}

\item{M}{integer, bootstrap replications}

\item{univ}{Boolean, perform univariate dimension reduction}

\item{rm_cross_terms}{Boolean, remove cross terms under univ}

\item{global_resids}{Boolean, use global residuals under univ}

\item{r}{integer, factor number}

\item{r_max}{integer, maximum factor number to consider}
}
\value{
mosumvar object
}
\description{
Change point detection for a factor model with VAR dynamics
}
\examples{
mosumvar_factor(panel$panel, p = 1, G = 24, method = "Score")
}
