% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcpt.R
\name{plot.factor_model}
\alias{plot.factor_model}
\title{Plot a factor_model object}
\usage{
\method{plot}{factor_model}(fm)
}
\arguments{
\item{fm}{factor_model object}
}
\value{
a plot of the information criterion against q, and a plot of the factor series
}
\description{
Plot a factor_model object
}
\examples{
fm <- factor_model(panel$panel)
plot(fm)
}
