% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosumvar_factor.R
\name{pooled_forecast}
\alias{pooled_forecast}
\title{Change point based forecasting methods for factor series}
\usage{
pooled_forecast(
  x,
  cp = 0,
  window_size = NULL,
  p = NULL,
  n.ahead = 1,
  weights = c("equal", "exp", "robust", "rolling")[1],
  return_model = F
)
}
\arguments{
\item{x}{matrix of data with series as columns}

\item{cp}{integer (vector) of estimated change points}

\item{window_size}{integer size of window to estimate back over}

\item{p}{integer VAR order}

\item{n.ahead}{integer number of steps ahead to forecast}

\item{weights}{String of weight method to use}

\item{return_model}{Boolean, return `ar` model}
}
\value{
list of class "pooled_forecast" containing matrix of forecasts, weight vector, and model
}
\description{
Change point based forecasting methods for factor series
}
\examples{
fm <- factor_model(panel$panel)
pooled_forecast(fm$f.q)
}
