% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcpt.R
\name{predict.factor_model}
\alias{predict.factor_model}
\title{Title}
\usage{
\method{predict}{factor_model}(fm, newdata)
}
\arguments{
\item{fm}{factor_model object}

\item{newdata}{matrix of (factor) data to predict from}
}
\value{
matrix of predictions
}
\description{
Title
}
