% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcpt.R
\name{summary.factor_model}
\alias{summary.factor_model}
\title{Title}
\usage{
\method{summary}{factor_model}(fm)
}
\arguments{
\item{fm}{factor_model object}
}
\value{
summary of factor_model object
}
\description{
Title
}
