% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcpt.R
\name{predict.factor_model}
\alias{predict.factor_model}
\title{Predict from a factor_model and forecasted factors}
\usage{
\method{predict}{factor_model}(fm, newdata)
}
\arguments{
\item{fm}{factor_model object}

\item{newdata}{matrix of (factor) data to predict from}
}
\value{
matrix of predictions
}
\description{
Predict from a factor_model and forecasted factors
}
\examples{
fm <- factor_model(panel$panel)
ar_fm <- ar(fm$f.q, p=1)
ar_pred <- predict(ar_fm, fm$f.q, n.ahead = 3)
fm_pred <- predict(fm, ar_pred$pred)
}
