% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorcpt.R
\name{summary.factor_model}
\alias{summary.factor_model}
\title{Summarise a factor_model object}
\usage{
\method{summary}{factor_model}(fm)
}
\arguments{
\item{fm}{factor_model object}
}
\value{
summary of factor_model object
}
\description{
Summarise a factor_model object
}
\examples{
fm <- factor_model(panel$panel)
summary(fm)
}
